CARMAGEDDON EDITION
This feed will be active for the week surrounding the upcoming I-405
closure, and will display links to information on Metro’s Free Fares and
Increased Service over the weekend. In addition some corrections were
made to the schedule for the Blue Line (Line 801), as well as minor
changes to two other routes (the 302 and the 705).